Added code * To verify that the posted ID is an INT and above 0

+++ del.php
@@ -9,6 +9,11 @@ if (DEBUG_MODE == true) {
 
 $tpl = new Template('templates/' . TEMPALTE);   // Creates the tpl object so we can reuse it
 $intFunctions = new internalFunctions;          // Creates the internalFunction object so we can call various functions (e.g. sending the header & footer)
+$ValidateInt = array(                           // Sets an option for the FILTER_VALIDATE_INT to allow anything above 0 and is an INT
+    'options' => array(
+        'min_range' => 0,
+    )
+);
 
 // Check if the user is logged in
 if (!checkedLoggedIn()){
@@ -23,8 +28,8 @@ if (!checkedLoggedIn()){
 
 $intFunctions->callHeader();                    // Call for the header
 
-if (isset($_GET['id'])) {
-    $id = mysqli_real_escape_string($dbcon, (int) $_GET['id']);
+if (isset($_GET['id']) && is_numeric($_GET['id'])) {
+    $id = mysqli_real_escape_string($dbcon, (int)filter_var($_GET['id'], FILTER_VALIDATE_INT, $ValidateInt));
     $sql = "DELETE FROM posts WHERE id = '$id'";
     $result = mysqli_query($dbcon, $sql);
     //$RowDeleted = mysqli_affected_rows($dbcon);
